Fix de tipos de datos en factura

+++ src/Presentation/ViewModels/FacturaVM/ActualizarFacturaViewModel.cs
@@ -6,7 +6,7 @@ namespace src.Presentation.ViewModels.FacturaVM
     {
         public string NumeroFactura { get; set; } = null!;
         public DateTime FechaEmision { get; set; }
-        public int IdCondicionPago { get; set; }
+        public short IdCondicionPago { get; set; }
         public bool Pagada { get; set; }
     }
 }
